Migrate to gorilla/mux for routing
